Change logic á»in repo file
