+++ apps/mobile/lib/features/authentication/domain/model/password.dart
