+++ apps/mobile/lib/features/authentication/domain/model/username.dart
