+++ apps/mobile/lib/features/authentication/domain/repo/src/authentication_repository.dart
@@ -1,29 +1,11 @@
-import 'dart:async';
-
-enum AuthenticationStatus { unknown, authenticated, unauthenticated }
+import '../../../../preferences/domain/model/user.dart';
+import 'authentication_logic.dart';
 
 class AuthenticationRepository {
-  final _controller = StreamController<AuthenticationStatus>();
+  final AuthenticationLogic _logic = AuthenticationLogic();
 
-  Stream<AuthenticationStatus> get status async* {
-    await Future<void>.delayed(const Duration(seconds: 1));
-    yield AuthenticationStatus.unauthenticated;
-    yield* _controller.stream;
-  }
+  Future<void> login(String username, String password) async =>
+      _logic.logIn(username: username, password: password);
 
-  Future<void> logIn({
-    required String username,
-    required String password,
-  }) async {
-    await Future.delayed(
-      const Duration(milliseconds: 300),
-      () => _controller.add(AuthenticationStatus.authenticated),
-    );
-  }
-
-  void logOut() {
-    _controller.add(AuthenticationStatus.unauthenticated);
-  }
-
-  void dispose() => _controller.close();
+  Future<void> logout() async => _logic.logOut();
 }
