+++ apps/mobile/lib/features/authentication/domain/use-cases/log_out/log_out_bloc.dart
@@ -1,0 +1,13 @@
+import 'package:bloc/bloc.dart';
+import 'package:meta/meta.dart';
+
+part 'log_out_event.dart';
+part 'log_out_state.dart';
+
+class LogOutBloc extends Bloc<LogOutEvent, LogOutState> {
+  LogOutBloc() : super(LogOutInitial()) {
+    on<LogOutEvent>((event, emit) {
+      // TODO: implement event handler
+    });
+  }
+}
