+++ apps/mobile/lib/features/authentication/domain/use-cases/log_out/log_out_event.dart
@@ -1,0 +1,4 @@
+part of 'log_out_bloc.dart';
+
+@immutable
+sealed class LogOutEvent {}
