+++ apps/mobile/lib/features/authentication/domain/use-cases/log_out/log_out_state.dart
@@ -1,0 +1,6 @@
+part of 'log_out_bloc.dart';
+
+@immutable
+sealed class LogOutState {}
+
+final class LogOutInitial extends LogOutState {}
