+++ apps/mobile/lib/features/authentication/domain/use-cases/sign_in/sign_in_bloc.dart
@@ -1,0 +1,13 @@
+import 'package:bloc/bloc.dart';
+import 'package:meta/meta.dart';
+
+part 'sign_in_event.dart';
+part 'sign_in_state.dart';
+
+class SignInBloc extends Bloc<SignInEvent, SignInState> {
+  SignInBloc() : super(SignInInitial()) {
+    on<SignInEvent>((event, emit) {
+      // TODO: implement event handler
+    });
+  }
+}
