+++ apps/mobile/lib/features/authentication/domain/use-cases/sign_in/sign_in_event.dart
@@ -1,0 +1,4 @@
+part of 'sign_in_bloc.dart';
+
+@immutable
+sealed class SignInEvent {}
