+++ apps/mobile/lib/features/authentication/domain/use-cases/sign_in/sign_in_state.dart
@@ -1,0 +1,6 @@
+part of 'sign_in_bloc.dart';
+
+@immutable
+sealed class SignInState {}
+
+final class SignInInitial extends SignInState {}
